+++ src/main/java/com/synopsys/integration/detect/lifecycle/boot/product/version/BlackDuckMinimumVersionChecks.java
@@ -1,4 +1,4 @@
-package com.synopsys.integration.detect.lifecycle.boot.product;
+package com.synopsys.integration.detect.lifecycle.boot.product.version;
 
 import java.util.LinkedList;
 import java.util.List;
@@ -8,30 +8,32 @@ import com.synopsys.integration.detect.configuration.enumeration.BlackduckScanMo
 import com.synopsys.integration.detect.configuration.enumeration.DetectTool;
 
 public class BlackDuckMinimumVersionChecks {
+    private final List<BlackDuckMinimumVersionCheck> checks;
 
-    public List<BlackDuckMinimumVersionCheck> create() {
-        List<BlackDuckMinimumVersionCheck> table = new LinkedList<>();
+    public BlackDuckMinimumVersionChecks() {
+        checks = new LinkedList<>();
 
-        table.add(new BlackDuckMinimumVersionCheck(
+        checks.add(new BlackDuckMinimumVersionCheck(
             "Rapid mode",
             c -> c.getValue(DetectProperties.DETECT_BLACKDUCK_SCAN_MODE) == BlackduckScanMode.RAPID,
             new BlackDuckVersion(2021, 6, 0)
         ));
 
-        table.add(new BlackDuckMinimumVersionCheck(
+        checks.add(new BlackDuckMinimumVersionCheck(
             "IaC scan",
             c -> c.getValue(DetectProperties.DETECT_TOOLS).containsValue(DetectTool.IAC_SCAN),
             new BlackDuckVersion(2021, 6, 0)
         ));
 
-        table.add(new BlackDuckMinimumVersionCheck(
+        checks.add(new BlackDuckMinimumVersionCheck(
             "Rapid signature scan",
             c -> c.getValue(DetectProperties.DETECT_TOOLS).containsValue(DetectTool.SIGNATURE_SCAN)
                 && c.getValue(DetectProperties.DETECT_BLACKDUCK_SCAN_MODE) == BlackduckScanMode.RAPID,
             new BlackDuckVersion(2022, 10, 0)
         ));
-
-        return table;
     }
 
+    public List<BlackDuckMinimumVersionCheck> get() {
+        return checks;
+    }
 }
